+++ example/src/screens/authenticated-user.tsx
@@ -1,5 +1,6 @@
 import React, { type ReactElement } from 'react';
 import { Text, TextInput, TouchableOpacity, StyleSheet, View, ScrollView, StatusBar } from 'react-native';
+import { logPageLoad } from 'react-native-ts-accountprotection';
 import type { MoneyTransferDTO } from '../App';
 
 interface AuthenticatedUserProps {
@@ -21,9 +22,13 @@ export class AuthenticatedUser extends React.Component<AuthenticatedUserProps, A
         super(props);
         this.state = {
             payerName: props.userId,
-            payeeName: '',
-            amount: '',
+            payeeName: 'Demo Payee',
+            amount: '100.00',
         };
+    }
+
+    componentDidMount() {
+        logPageLoad('AuthenticatedUserScreen');
     }
 
     render() {